[Ignition]: Allow for adapter specific data passed by the adapter when a position is opened.
In preparation to support the calculation of fees, this commit makes it so that the adapters are able to pass some opaque data object to the protocol with the expectation that the protocol will pass them this same data object back again when the liquidity position is being closed. This can be used for many things, but the primary usage of this is to allow for the calculation of fees by the adapters.